CC_thermodynamics: adding sat adj using all double variables; CloudClousre_s: adding computations from thetali-input; test_CC: update prints

+++ parameters.py
@@ -1,27 +1,29 @@
+import numpy as np
+
 # //Generated code! Absolutely DO NOT modify this file, parameters should be modified in generate_parameters.py
 def pi(): 3.14159265359
 global g
 g = 9.80665
 global Rd, Rv
-Rd = 287.1
-Rv = 461.5
+Rd = np.double(287.1)
+Rv = np.double(461.5)
 global eps_v, eps_vi
-eps_v = 0.6221018418201517
-eps_vi = 1.6074538488331591
+eps_v = np.double(0.6221018418201517)
+eps_vi = np.double(1.6074538488331591)
 global cpd, cpv, cl, ci
-cpd = 1004.0
-cpv = 1859.0
-cl = 4218.0
-ci = 2106.0
-kappa = 0.285956175299
+cpd = np.double(1004.0)
+cpv = np.double(1859.0)
+cl = np.double(4218.0)
+ci = np.double(2106.0)
+kappa = np.double(0.285956175299)
 # define Tf 273.15
 # define Tt 273.16
 global T_tilde, p_tilde, sd_tilde, sv_tilde
-T_tilde = 298.15
-p_tilde = 100000.0
+T_tilde = np.double(298.15)
+p_tilde = np.double(100000.0)
 # define pv_star_t 611.7
-sd_tilde = 6864.8
-sv_tilde = 10513.6
+sd_tilde = np.double(6864.8)
+sv_tilde = np.double(10513.6)
 # define omega 7.29211514671e-05
 # define ql_threshold 1e-08
 # define vkb 0.35
